the end is near

+++ ClassLibrary1/Rectangle.cs
@@ -24,7 +24,7 @@ namespace FiguresClassLibrary
             }
         }
 
-        public Rectangle(int x1, int y1, int x2, int y2, int x3, int y3, int x4, int y4)
+        public Rectangle(double x1, double y1, double x2, double y2, double x3, double y3, double x4, double y4)
         {
             double a = Math.Sqrt(Math.Pow(x2 - x1, 2) + Math.Pow(y2 - y1, 2));
             double b = Math.Sqrt(Math.Pow(x3 - x2, 2) + Math.Pow(y3 - y2, 2));
