+++ ClassLibrary1/Triangle.cs
@@ -10,7 +10,7 @@ namespace FiguresClassLibrary
 
         public Triangle(double a, double b, double c)
         {
-            if( a + b > c && a + c > b && b + c > a)
+            if( a + b > c && a + c > b && b + c > a )
             {
                 _a = a;
                 _b = b;
@@ -22,13 +22,13 @@ namespace FiguresClassLibrary
             }
         }
 
-        public Triangle(int x1, int y1, int x2, int y2, int x3, int y3)
+        public Triangle(double x1, double y1, double x2, double y2, double x3, double y3)
         {
             double a = Math.Sqrt(Math.Pow(x2 - x1, 2) + Math.Pow(y2 - y1, 2));
             double b = Math.Sqrt(Math.Pow(x3 - x2, 2) + Math.Pow(y3 - y2, 2));
             double c = Math.Sqrt(Math.Pow(x1 - x3, 2) + Math.Pow(y1 - y3, 2));
 
-            if (a + b > c && a + c > b && b + c > a)
+            if (a + b > c && a + c > b && b + c > a )
             {
                 _a = a;
                 _b = b;
